refactor: 💡 optimize code orgnization
